Add KR_ENABLE_FILESYSTEM; load PEM from filename
Allow certificate and key objects to be passed directly
in the argument. PEM objects are distinctive enough to not be
confused with valid filenames, so this support remains enabled even when
filesystem is enabled, but is mainly intended to be used when it's not.

PUBLISHED_FROM=a9c783a824187a820aa7e0d8a198bdd4b0ad914a